careerhive/maf-service/requirements.txt
semantic-kernel==1.17.1
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.27.2
